vkd3d: Fix formatting warnings.
Extremely annoying and there's a reason why we'd had -Wno-format,
but there were tons of (mostly benign) formatting bugs all over the place.

Native and MinGW builds use different types for things mostly due to
long. There seems to be no good way to have MinGW assume that %zu works since it still
targets msvcrt.dll from prehistoric era, so needed a hack to get around that too.

+++ libs/d3d12core/debug.c
@@ -54,7 +54,7 @@ static HRESULT STDMETHODCALLTYPE d3d12_dred_settings_QueryInterface(
 static ULONG STDMETHODCALLTYPE d3d12_dred_settings_AddRef(d3d12_dred_settings_iface *iface)
 {
     struct d3d12_dred_settings *dred_settings = impl_from_ID3D12DeviceRemovedExtendedDataSettings(iface);
-    ULONG refcount = InterlockedIncrement(&dred_settings->refcount);
+    unsigned int refcount = InterlockedIncrement(&dred_settings->refcount);
 
     TRACE("%p increasing refcount to %u.\n", dred_settings, refcount);
 
@@ -64,7 +64,7 @@ static ULONG STDMETHODCALLTYPE d3d12_dred_settings_AddRef(d3d12_dred_settings_if
 static ULONG STDMETHODCALLTYPE d3d12_dred_settings_Release(d3d12_dred_settings_iface *iface)
 {
     struct d3d12_dred_settings *dred_settings = impl_from_ID3D12DeviceRemovedExtendedDataSettings(iface);
-    ULONG refcount;
+    unsigned int refcount;
 
     pthread_mutex_lock(&debug_singleton_lock);
     refcount = InterlockedDecrement(&dred_settings->refcount);
